@@ -1,5 +1,5 @@
 {
-  "name": "udagram-api",
+  "name": "udagram-api-feed",
   "version": "2.0.0",
   "description": "",
   "main": "src/server.js",
@@ -45,6 +45,6 @@
     "eslint-config-google": "^0.14.0",
     "mocha": "^6.1.4",
     "ts-node-dev": "^1.0.0-pre.32",
-    "typescript": "^3.3.4000"
+    "typescript": "^4.9.5"
   }
 }
