@@ -1,5 +1,5 @@
 {
-  "name": "udagram-api",
+  "name": "udagram-api-user",
   "version": "2.0.0",
   "description": "",
   "main": "src/server.js",
@@ -18,6 +18,7 @@
   "dependencies": {
     "@types/bcrypt": "^3.0.0",
     "@types/jsonwebtoken": "^8.3.2",
+    "@types/lodash": "^4.14.199",
     "aws-sdk": "^2.429.0",
     "bcrypt": "^5.0.1",
     "body-parser": "^1.18.3",
@@ -33,7 +34,7 @@
   "devDependencies": {
     "@types/bluebird": "^3.5.26",
     "@types/cors": "^2.8.6",
-    "@types/express": "^4.16.1",
+    "@types/express": "^4.17.18",
     "@types/node": "^11.11.6",
     "@types/sequelize": "^4.27.44",
     "@types/validator": "^10.9.0",
@@ -45,6 +46,6 @@
     "eslint-config-google": "^0.14.0",
     "mocha": "^6.1.4",
     "ts-node-dev": "^1.0.0-pre.32",
-    "typescript": "^3.3.4000"
+    "typescript": "^4.9.5"
   }
 }
