@@ -1,22 +1,23 @@
-import AWS = require('aws-sdk');
-import {config} from './config/config';
-
+import AWS = require("aws-sdk");
+import {config} from "./config/config";
 
 // Configure AWS
-const credentials = new AWS.SharedIniFileCredentials({profile: config.aws_profile});
+const credentials = new AWS.SharedIniFileCredentials({
+  profile: config.aws_profile,
+});
 AWS.config.credentials = credentials;
 
 export const s3 = new AWS.S3({
-  signatureVersion: 'v4',
+  signatureVersion: "v4",
   region: config.aws_region,
   params: {Bucket: config.aws_media_bucket},
 });
 
 // Generates an AWS signed URL for retrieving objects
-export function getGetSignedUrl( key: string ): string {
+export function getGetSignedUrl(key: string): string {
   const signedUrlExpireSeconds = 60 * 5;
 
-  return s3.getSignedUrl('getObject', {
+  return s3.getSignedUrl("getObject", {
     Bucket: config.aws_media_bucket,
     Key: key,
     Expires: signedUrlExpireSeconds,
@@ -24,10 +25,10 @@ export function getGetSignedUrl( key: string ): string {
 }
 
 // Generates an AWS signed URL for uploading objects
-export function getPutSignedUrl( key: string ): string {
+export function getPutSignedUrl(key: string): string {
   const signedUrlExpireSeconds = 60 * 5;
 
-  return s3.getSignedUrl('putObject', {
+  return s3.getSignedUrl("putObject", {
     Bucket: config.aws_media_bucket,
     Key: key,
     Expires: signedUrlExpireSeconds,
