@@ -1,17 +1,19 @@
-import {Sequelize} from 'sequelize-typescript';
-import {config} from './config/config';
+import {Sequelize} from "sequelize-typescript";
+import {config} from "./config/config";
+
+const configSequelize: any = config;
 
 export const sequelize = new Sequelize({
-  'username': config.username,
-  'password': config.password,
-  'database': config.database,
-  'host': config.host,
-  'dialect': config.dialect,
-  'storage': ':memory:',
-  'dialectOptions': {
+  username: configSequelize.username,
+  password: configSequelize.password,
+  database: configSequelize.database,
+  host: configSequelize.host,
+  dialect: configSequelize.dialect,
+  storage: ":memory:",
+  dialectOptions: {
     ssl: {
       require: true,
-      rejectUnauthorized: false
-    }
-  }
+      rejectUnauthorized: false,
+    },
+  },
 });
