@@ -1,10 +1,12 @@
 import cors from "cors";
 import express from "express";
-import { sequelize } from "./sequelize";
-import { IndexRouter } from "./controllers/v0/index.router";
+import {sequelize as sequelizeObj} from "./sequelize";
+import {IndexRouter} from "./controllers/v0/index.router";
 import bodyParser from "body-parser";
-import { config } from "./config/config";
-import { V0_USER_MODELS } from "./controllers/v0/model.index";
+import {config} from "./config/config";
+import {V0_USER_MODELS} from "./controllers/v0/model.index";
+
+const sequelize: any = sequelizeObj;
 
 (async () => {
   await sequelize.addModels(V0_USER_MODELS);
